backend/requirements.txt
fastapi
pandas
sqlalchemy
psycopg2-binary
uvicorn
pydantic
pydantic[email]
passlib
pyjwt
python-dotenv
pytest
langchain
langgraph
langchain-openai
langchain-groq
langchain-ollama
langchain-community
autopep8
prettier
coloredlogs